requirements.txt
python-nmap
requests
gevent